+++ src/test/java/subway/StationAcceptanceTest.java
@@ -149,57 +149,4 @@ public class StationAcceptanceTest {
         assertThat(stationNames).doesNotContain("종합운동장");
         assertThat(stationNames).size().isEqualTo(0);
     }
-
-    @DisplayName("지하철 노선을 생성하여 새로운 노선을 추가한다.")
-    @Test
-    public void testCreateLine(){
-        //도메인 제약 테스트를 추가하기 위해 노선의 조건에서 색상이 중복되지 않고 거리가 100이하여야 한다고 가정한다
-
-        //given 역을 생성하고
-        ExtractableResponse<Response> createStationResponse = RestAssured.given().log().all()
-          .body(new StationRequest("종합운동장"))
-          .contentType(MediaType.APPLICATION_JSON_VALUE)
-          .when().post("/stations")
-          .then().log().all()
-          .extract();
-
-        ExtractableResponse<Response> createStationResponse2 = RestAssured.given().log().all()
-          .body(new StationRequest("잠실새내"))
-          .contentType(MediaType.APPLICATION_JSON_VALUE)
-          .when().post("/stations")
-          .then().log().all()
-          .extract();
-
-        //when 두개의 역에 대한 지하철 노선을 추가한다.
-        ExtractableResponse<Response> createLineResponse = RestAssured.given().log().all()
-          .body(new LineCreateRequest("2호선", "green", 1L, 2L, 10))
-          .contentType(MediaType.APPLICATION_JSON_VALUE)
-          .when().post("/lines")
-          .then().log().all()
-          .extract();
-
-        //then
-        // 이미 노선이 있는지 확인한다.
-        ExtractableResponse<Response> showResponse = RestAssured.given().log().all()
-          .contentType(MediaType.APPLICATION_JSON_VALUE)
-          .when().get("/lines")
-          .then().log().all()
-          .extract();
-
-        assertThat(showResponse.body().jsonPath().getList("name", String.class)).size().isEqualTo(1);
-
-        // 노선이 있지 않다면 두개의 역에 대한 지하철 노선이 추가되었는지 확인한다.
-        assertThat(createStationResponse.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(createLineResponse.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(createLineResponse.body().jsonPath().getList("stations", Long.class)).size().isEqualTo(2);
-        assertThat(createLineResponse.body().jsonPath().getList("stations", Long.class).get(0)).isEqualTo(1L);
-        assertThat(createLineResponse.body().jsonPath().getList("stations", Long.class).get(1)).isEqualTo(2L);
-        //첫번째 이름 종합운동장인지, 두번째 이름 잠실새내역인지 확인
-        assertThat(createLineResponse.body().jsonPath().getString("stations[0].name")).isEqualTo("종합운동장");
-        assertThat(createLineResponse.body().jsonPath().getString("stations[1].name")).isEqualTo("잠실새내");
-        assertThat(createLineResponse.body().jsonPath().getString("name")).isEqualTo("2호선");
-        assertThat(createLineResponse.body().jsonPath().getString("color")).isEqualTo("green");
-        assertThat(createLineResponse.body().jsonPath().getLong("distance")).isEqualTo(10);
-
-    }
 }
